Modify copying in tmux
Copying in tmux is now done by first pressing the leader key (Ctrl+g at
the time of writing), before entering vi copy mode with Ctrl+c. It is
then possible to navigate the tmux pane with the vim-movement keys hjkl.
Selection is started with 'v' (or Ctrl+v for rectangular selection mode)
and the content of the selection is yanked with 'y'. The selection is
then copied to the system clipboard.